% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decode_links}
\alias{decode_links}
\title{Decode short links}
\usage{
decode_links(tbl)
}
\arguments{
\item{tbl}{timeline extracted with \code{rtweet}}
}
\value{
tibble of urls
}
\description{
Simple convenience function to check for folder and create if not found.
}
\examples{
\dontrun{
decode_links(tbl)
}
}
