% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_create.R
\name{dir_create}
\alias{dir_create}
\title{Create new directory}
\usage{
dir_create(x)
}
\arguments{
\item{x}{file path}
}
\value{

}
\description{
Simple convenience function to check for folder and create if not found.
}
\examples{
\dontrun{
dir_create("~/Desktop/New_Folder")
}
}
