% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pics.R
\name{get_pics}
\alias{get_pics}
\title{Clean Twitter timeline and get images}
\usage{
get_pics(dir, tbl)
}
\arguments{
\item{dir}{directory to save files into (predetermined)}

\item{tbl}{twitter timeline in tibble format}
}
\value{
tibble with image data
}
\description{
Take a tibble from \code{rtweet::get_timeline()}, extract urls, order, and
read in images
}
\examples{
\dontrun{
pics <- get_pics(d2, tl)
}
}
