% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_read_fn.R
\name{image_read_fn}
\alias{image_read_fn}
\title{Read and scale images}
\usage{
image_read_fn(img)
}
\arguments{
\item{img}{url or location of image to read in}
}
\value{

}
\description{
Read in an image, scale it, and set white border. Just streamlines the \code{magick::image_read()} function.
}
\examples{
\dontrun{
image_read_fn("https://jeroen.github.io/images/frink.png")
}
}
