% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_collage.R
\name{make_blanks}
\alias{make_blanks}
\title{if images are missing or column count is off, build blanks}
\usage{
make_blanks(bf, year)
}
\arguments{
\item{bf}{file name, automatically created}

\item{year}{year for file save}
}
\description{
if images are missing or column count is off, build blanks
}
