% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_collage.R
\name{make_collage}
\alias{make_collage}
\title{Make a collage of images from Twitter timeline}
\usage{
make_collage(year)
}
\arguments{
\item{year}{vector of years to compile}
}
\value{
12xn collage in /final_cols/ folder
}
\description{
Create a 12xN size collage. Takes photos and appends them into
column, then stitches columns together into collage. Collages are split out
by the year images were posted. To ensure proper sizing, blank cells are
created to fill out any NA or missing images.
}
\examples{
\dontrun{
years \%>\% purrr::walk(make_collage)
}
}
