% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_collage.R
\name{make_column}
\alias{make_column}
\title{build the image columns to be collaged}
\usage{
make_column(i, files, no_rows, year)
}
\arguments{
\item{i}{number of columns}

\item{files}{vector of image files that were downloaded}

\item{no_rows}{how many rows in each column}

\item{year}{year images posted}
}
\description{
build the image columns to be collaged
}
