% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_btm.R
\name{run_btm}
\alias{run_btm}
\title{Run Biterm Topic Models}
\usage{
run_btm(handle, data_source = c("timeline", "mentions"), n_topics, n_terms)
}
\arguments{
\item{handle}{twitter handle without the \code{@} sign}

\item{data_source}{either timeline or mentions}

\item{n_topics}{how many topics to model}

\item{n_terms}{how many of the most salient terms to include for each topic}
}
\value{
.rds file - x_phi is scaled x,y coordinates for topics, term_info is the top n terms;
both are based on lambda - 0.6; json is the json object from \code{LDAvis} package
}
\description{
Uses the \code{BTM} package to run noun-based topic models in collections of short texts.
Uses \code{udpipe} to tokenize and keep only nouns, proper nouns. The BTM output is then processed using
functions and calculations from the \code{LDAvis} package.
}
\details{
The json file is included in the output if you want to visualize with \code{LDAvis}
or plots can be made with the x_phi output.
}
\examples{
\dontrun{
run_btm("BMWUSA", data_source = 'timeline', n_topics = 40, n_terms = 30)
}
}
