% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_tm.R
\name{twitter_tm}
\alias{twitter_tm}
\title{Get Twitter Timeline, Mentions and Topic Models}
\usage{
twitter_tm(handle, search_query = NULL, topic_ct, term_ct)
}
\arguments{
\item{handle}{Twitter handle without \code{@} sign}

\item{search_query}{boolean search query}

\item{topic_ct}{how many topics to model}

\item{term_ct}{how many top salient terms to return for each topic}
}
\value{

}
\description{
Total convenience function to extract Twitter timeline of specified handle,
18,000 tweets mentioning specific search query, and then uses biterm topic model
algorithm to model TM of both the timeline and mentions.
}
\details{
This is the only function that needs to be run. Everything else works in the background.

Relies heavily on \code{rtweet} \code{BTM} and \code{LDAvis} packages.
}
\examples{
\dontrun{
twitter_tm("BMWUSA", search_query = "bmw", topic_ct = 40, term_ct = 30)
}
}
