% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_twitter.R
\name{get_twitter}
\alias{get_twitter}
\title{Get Twitter Timeline}
\usage{
get_twitter(handle, search_query = NULL)
}
\arguments{
\item{handle}{twitter handle without \code{@}}

\item{search_query}{boolean search query (read \code{rtweet::search_tweets()} docs for proper formatting)}
}
\value{

}
\description{
Extract 3,200 tweets from a timeline, extract all pictures and save timeline
asa .rds file to new folder. Note that this creates a new directory in the current
working directory. "~/working_directory/account_analysis/handle/Sys.Date"
}
\details{
If search_query is added, also pulls 18,000 recent tweets matching the query.
Mentions are also saved into the .rds file.

Heavily reliant on the \code{rtweet} package. Read the package tutorial for how to
set up Twitter API authorization.
}
\examples{
\dontrun{
get_twitter("BMWUSA", search_query = "bmw")
}
}
